feat: apply facade pattern to 3 additional large files
- safe-events-manager.ts: 1,741 → 485 lines (72.1% reduction)
- data-service-adapter.ts: 1,716 → 485 lines (71.7% reduction)
- composite-system.ts: 1,685 → 475 lines (71.8% reduction)

All facades delegate to specialized @claude-zen packages while maintaining API compatibility:
- SAFe events management → @claude-zen/safe-framework
- Multi-database operations → @claude-zen/database + @claude-zen/foundation
- Agent coordination → @claude-zen/teamwork + @claude-zen/load-balancing + @claude-zen/brain

Total cumulative reduction: 11,227 → 2,612 lines (76.7% overall reduction)

+++ packages/brain/rust/core/cascade.rs
@@ -950,7 +950,7 @@ impl<T: Float> CascadeTrainer<T> {
         let expected_outputs = network.num_outputs();
 
         for (i, input) in data.inputs.iter().enumerate() {
-            if input.len() != expected_inputs {
+            if input.len() != expected_inputs as usize {
                 return Err(CascadeError::InvalidConfiguration(format!(
                     "Input {} has wrong size: expected {}, got {}",
                     i,
@@ -961,7 +961,7 @@ impl<T: Float> CascadeTrainer<T> {
         }
 
         for (i, output) in data.outputs.iter().enumerate() {
-            if output.len() != expected_outputs {
+            if output.len() != expected_outputs as usize {
                 return Err(CascadeError::InvalidConfiguration(format!(
                     "Output {} has wrong size: expected {}, got {}",
                     i,
